+++ crates/ic-query/src/sns/report/source/model/mod.rs
@@ -14,6 +14,6 @@ pub(in crate::sns::report) use fetch::SnsFetchRequest;
 pub use fetch::SnsSourceRequest;
 pub(in crate::sns::report) use list::MainnetSnsCanisters;
 pub use list::{MainnetSns, MainnetSnsList};
-pub(in crate::sns::report) use neurons::{MainnetSnsNeuronPage, MainnetSnsNeurons, SnsNeuronId};
+pub use neurons::{MainnetSnsNeuronPage, MainnetSnsNeurons, SnsNeuronId};
 pub use proposals::{MainnetSnsProposal, MainnetSnsProposalPage, MainnetSnsProposals};
 pub use token::MainnetSnsToken;